Crate rusoto_ram

Use AWS Resource Access Manager to share AWS resources between AWS accounts. To share a resource, you create a resource share, associate the resource with the resource share, and specify the principals that can access the resource. The following principals are supported:

The ID of an AWS account

The Amazon Resource Name (ARN) of an OU from AWS Organizations

The Amazon Resource Name (ARN) of an organization from AWS Organizations

If you specify an AWS account that doesn't exist in the same organization as the account that owns the resource share, the owner of the specified account receives an invitation to accept the resource share. After the owner accepts the invitation, they can access the resources in the resource share. An administrator of the specified account can use IAM policies to restrict access resources in the resource share.

If you're using the service, you're probably looking for RamClient and Ram.

Structs

AcceptResourceShareInvitationRequest

AcceptResourceShareInvitationResponse

AssociateResourceShareRequest

AssociateResourceShareResponse

CreateResourceShareRequest

CreateResourceShareResponse

DeleteResourceShareRequest

DeleteResourceShareResponse

DisassociateResourceShareRequest

DisassociateResourceShareResponse

EnableSharingWithAwsOrganizationRequest

EnableSharingWithAwsOrganizationResponse

GetResourcePoliciesRequest

GetResourcePoliciesResponse

GetResourceShareAssociationsRequest

GetResourceShareAssociationsResponse

GetResourceShareInvitationsRequest

GetResourceShareInvitationsResponse

GetResourceSharesRequest

GetResourceSharesResponse

ListPrincipalsRequest

ListPrincipalsResponse

ListResourcesRequest

ListResourcesResponse

Principal
Describes a principal for use with AWS Resource Access Manager.

RamClient
A client for the RAM API.

RejectResourceShareInvitationRequest

RejectResourceShareInvitationResponse

Resource
Describes a resource associated with a resource share.

ResourceShare
Describes a resource share.

ResourceShareAssociation
Describes an association with a resource share.

ResourceShareInvitation
Describes an invitation to join a resource share.

Tag
Information about a tag.

TagFilter
Used to filter information based on tags.

TagResourceRequest

TagResourceResponse

UntagResourceRequest

UntagResourceResponse

UpdateResourceShareRequest

UpdateResourceShareResponse

Enums

AcceptResourceShareInvitationError
Errors returned by AcceptResourceShareInvitation

AssociateResourceShareError
Errors returned by AssociateResourceShare

CreateResourceShareError
Errors returned by CreateResourceShare

DeleteResourceShareError
Errors returned by DeleteResourceShare

DisassociateResourceShareError
Errors returned by DisassociateResourceShare

EnableSharingWithAwsOrganizationError
Errors returned by EnableSharingWithAwsOrganization

GetResourcePoliciesError
Errors returned by GetResourcePolicies

GetResourceShareAssociationsError
Errors returned by GetResourceShareAssociations

GetResourceShareInvitationsError
Errors returned by GetResourceShareInvitations

GetResourceSharesError
Errors returned by GetResourceShares

ListPrincipalsError
Errors returned by ListPrincipals

ListResourcesError
Errors returned by ListResources

RejectResourceShareInvitationError
Errors returned by RejectResourceShareInvitation

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateResourceShareError
Errors returned by UpdateResourceShare

Traits

Ram
Trait representing the capabilities of the RAM API. RAM clients implement this trait.