Struct rusoto_ram::RamClient

pub struct RamClient { /* fields omitted */ }

A client for the RAM API.

Methods

impl RamClient

pub fn new(region: Region) -> RamClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> RamClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,