Trait rusoto_ram::Ram

pub trait Ram {
    fn accept_resource_share_invitation(
        &self, 
        input: AcceptResourceShareInvitationRequest
    ) -> RusotoFuture<AcceptResourceShareInvitationResponse, AcceptResourceShareInvitationError>;

    fn associate_resource_share(
        &self, 
        input: AssociateResourceShareRequest
    ) -> RusotoFuture<AssociateResourceShareResponse, AssociateResourceShareError>;

    fn create_resource_share(
        &self, 
        input: CreateResourceShareRequest
    ) -> RusotoFuture<CreateResourceShareResponse, CreateResourceShareError>;

    fn delete_resource_share(
        &self, 
        input: DeleteResourceShareRequest
    ) -> RusotoFuture<DeleteResourceShareResponse, DeleteResourceShareError>;

    fn disassociate_resource_share(
        &self, 
        input: DisassociateResourceShareRequest
    ) -> RusotoFuture<DisassociateResourceShareResponse, DisassociateResourceShareError>;

    fn enable_sharing_with_aws_organization(
        &self
    ) -> RusotoFuture<EnableSharingWithAwsOrganizationResponse, EnableSharingWithAwsOrganizationError>;

    fn get_resource_policies(
        &self, 
        input: GetResourcePoliciesRequest
    ) -> RusotoFuture<GetResourcePoliciesResponse, GetResourcePoliciesError>;

    fn get_resource_share_associations(
        &self, 
        input: GetResourceShareAssociationsRequest
    ) -> RusotoFuture<GetResourceShareAssociationsResponse, GetResourceShareAssociationsError>;

    fn get_resource_share_invitations(
        &self, 
        input: GetResourceShareInvitationsRequest
    ) -> RusotoFuture<GetResourceShareInvitationsResponse, GetResourceShareInvitationsError>;

    fn get_resource_shares(
        &self, 
        input: GetResourceSharesRequest
    ) -> RusotoFuture<GetResourceSharesResponse, GetResourceSharesError>;

    fn list_principals(
        &self, 
        input: ListPrincipalsRequest
    ) -> RusotoFuture<ListPrincipalsResponse, ListPrincipalsError>;

    fn list_resources(
        &self, 
        input: ListResourcesRequest
    ) -> RusotoFuture<ListResourcesResponse, ListResourcesError>;

    fn reject_resource_share_invitation(
        &self, 
        input: RejectResourceShareInvitationRequest
    ) -> RusotoFuture<RejectResourceShareInvitationResponse, RejectResourceShareInvitationError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;

    fn update_resource_share(
        &self, 
        input: UpdateResourceShareRequest
    ) -> RusotoFuture<UpdateResourceShareResponse, UpdateResourceShareError>;
}

Trait representing the capabilities of the RAM API. RAM clients implement this trait.

Required methods

fn accept_resource_share_invitation(
&self,
input: AcceptResourceShareInvitationRequest
) -> RusotoFuture<AcceptResourceShareInvitationResponse, AcceptResourceShareInvitationError>

Accepts an invitation to a resource share from another AWS account.

fn associate_resource_share(
&self,
input: AssociateResourceShareRequest
) -> RusotoFuture<AssociateResourceShareResponse, AssociateResourceShareError>

Associates the specified resource share with the specified principals and resources.

fn create_resource_share(
&self,
input: CreateResourceShareRequest
) -> RusotoFuture<CreateResourceShareResponse, CreateResourceShareError>

Creates a resource share.

fn delete_resource_share(
&self,
input: DeleteResourceShareRequest
) -> RusotoFuture<DeleteResourceShareResponse, DeleteResourceShareError>

Deletes the specified resource share.

fn disassociate_resource_share(
&self,
input: DisassociateResourceShareRequest
) -> RusotoFuture<DisassociateResourceShareResponse, DisassociateResourceShareError>

Disassociates the specified principals or resources from the specified resource share.

fn enable_sharing_with_aws_organization(
&self
) -> RusotoFuture<EnableSharingWithAwsOrganizationResponse, EnableSharingWithAwsOrganizationError>

Enables resource sharing within your organization.

fn get_resource_policies(
&self,
input: GetResourcePoliciesRequest
) -> RusotoFuture<GetResourcePoliciesResponse, GetResourcePoliciesError>

Gets the policies for the specifies resources.

fn get_resource_share_associations(
&self,
input: GetResourceShareAssociationsRequest
) -> RusotoFuture<GetResourceShareAssociationsResponse, GetResourceShareAssociationsError>

Gets the associations for the specified resource share.

fn get_resource_share_invitations(
&self,
input: GetResourceShareInvitationsRequest
) -> RusotoFuture<GetResourceShareInvitationsResponse, GetResourceShareInvitationsError>

Gets the specified invitations for resource sharing.

fn get_resource_shares(
&self,
input: GetResourceSharesRequest
) -> RusotoFuture<GetResourceSharesResponse, GetResourceSharesError>

Gets the specified resource shares or all of your resource shares.

fn list_principals(
&self,
input: ListPrincipalsRequest
) -> RusotoFuture<ListPrincipalsResponse, ListPrincipalsError>

Lists the principals with access to the specified resource.

fn list_resources(
&self,
input: ListResourcesRequest
) -> RusotoFuture<ListResourcesResponse, ListResourcesError>

Lists the resources that the specified principal can access.

fn reject_resource_share_invitation(
&self,
input: RejectResourceShareInvitationRequest
) -> RusotoFuture<RejectResourceShareInvitationResponse, RejectResourceShareInvitationError>

Rejects an invitation to a resource share from another AWS account.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Adds the specified tags to the specified resource share.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Removes the specified tags from the specified resource share.

fn update_resource_share(
&self,
input: UpdateResourceShareRequest
) -> RusotoFuture<UpdateResourceShareResponse, UpdateResourceShareError>

Updates the specified resource share.

Loading content...